Move PwaApplication, repo and service to the domain package. Huge number of files affected indicate to me this is probably the right place.
Looking ahead, PwaApplicationDetail is probably just a feature and not the domain, although that reorg will probably touch just as many or more files. "Detail versions" seems to me to be an application feature, not a core concept, because you probably wouldn't use it as a term when discussing the core PWA system.