Merged in hotfix/add-batching-to-pipeline-ident-view-construction (pull request #1102)
Add batching to ident view collector processing to we avoid issues with the Oracle IN clause 1000 item limit.

* Filter pad pipelines bvy application detail when constructing pipeline and ident views.

This is a short term fix for an issue where an application with a large number of pipelines and idents blew out the 1000 limit on oracle SQL IN clause.

The issue may recur in exceptionally large apps where the total legitimate number of idents for the app pipelines exceeds 1000. This could happen if an application contained 200 pipelines, each with 5 or more idents.

* Add batching to ident view collector processing to we avoid issues with the Oracle IN clause 1000 item limit.
Add extra comments and helper debug statements.

* Fix comment tYp0.

* Merged hotfix/5 into hotfix/add-batching-to-pipeline-ident-view-construction
* Merged hotfix/5 into hotfix/add-batching-to-pipeline-ident-view-construction

Approved-by: Chris Wood